+++ passes/S043/S043_test.go
@@ -1,0 +1,14 @@
+package S043_test
+
+import (
+	"testing"
+
+	"golang.org/x/tools/go/analysis/analysistest"
+
+	"github.com/jfrappier/tfsprout/passes/S043"
+)
+
+func TestS043(t *testing.T) {
+	testdata := analysistest.TestData()
+	analysistest.Run(t, testdata, S043.Analyzer, "testdata/src/a")
+}